4282: authorization

+++ app/controllers/admin/project_logs_controller.rb
@@ -4,6 +4,10 @@ class Admin::ProjectLogsController < Admin::AdminController
     @step = ProjectStep.find(@log.project_step_id)
     @loan = Loan.find(@step.project_id)
 
+    authorize @log
+    authorize @step
+    authorize @loan
+
     if @log.destroy
       redirect_to admin_loan_path(@loan, anchor: 'timeline'), notice: I18n.t(:notice_deleted)
     else
@@ -15,6 +19,10 @@ class Admin::ProjectLogsController < Admin::AdminController
     @log = ProjectLog.find(params[:id])
     @step = ProjectStep.find(@log.project_step_id)
     @loan = Loan.find(@step.project_id)
+
+    authorize @log
+    authorize @step
+    authorize @loan
 
     redirect_to admin_loan_path(@loan)
   end
